Frontend: Refactor: Extract Interface into own class
