+++ apps/frontend/app/app/(app)/account-balance/MyCardReader.ts
@@ -1,20 +1,14 @@
-import CardResponse from '@/helper/nfcCardReaderHelper/CardResponse';
-import { isRunningInExpoGo } from 'expo';
-import { Platform } from 'react-native';
+import {isRunningInExpoGo} from 'expo';
+import {Platform} from 'react-native';
 import MyNativeCardReader from './MyNativeCardReader';
 import MyUnsupportedCardReader from './MyUnsupportedCardReader';
+import {MyCardReaderInterface} from "@/app/(app)/account-balance/MyCardReaderInterface";
 
 export type MyCardReaderResponseSupport = {
 	result: boolean;
 	message?: string;
 	error?: any;
 };
-
-export interface MyCardReaderInterface {
-	isNfcSupported: () => Promise<MyCardReaderResponseSupport>;
-	isNfcEnabled: () => Promise<MyCardReaderResponseSupport>;
-	readCard: (callBack: (answer: CardResponse | undefined) => Promise<void>, showInstruction: () => void, hideInstruction: () => void, nfcInstruction: string) => Promise<void>;
-}
 
 export default function useMyCardReader(): MyCardReaderInterface {
 	const isExpoGo = isRunningInExpoGo();
