+++ apps/frontend/app/app/(app)/account-balance/MyCardReaderInterface.ts
@@ -1,0 +1,13 @@
+import CardResponse from '@/helper/nfcCardReaderHelper/CardResponse';
+
+export type MyCardReaderResponseSupport = {
+	result: boolean;
+	message?: string;
+	error?: any;
+};
+
+export interface MyCardReaderInterface {
+	isNfcSupported: () => Promise<MyCardReaderResponseSupport>;
+	isNfcEnabled: () => Promise<MyCardReaderResponseSupport>;
+	readCard: (callBack: (answer: CardResponse | undefined) => Promise<void>, showInstruction: () => void, hideInstruction: () => void, nfcInstruction: string) => Promise<void>;
+}
